fix: rename border-swtich -> border-switch
