+++ src/renderer/src/components/base/border-switch.tsx
